NovaPlan_DEMO_v0.1
This is a demo of the AI roadmap generator that helps users that don't know what field they should study/work in based on their personal summary and on their answers to a few questions. The roadmaps are detailed and they contain milestones, tasks, scholarships, courses, and the duration to complete said milestones.

The demo contains a few fields. The full version of the AI will contain the entirety of the O*NET and ESCO datasets.